Fix the posts disappearing bug.
In 368a0101 an `Array.forEach` was refactored into a normal `for..of`
loop, but an early out `return` to skip adding quotes for in-batch
posts wasn't changed to a `continue`. Thus, if an update contained
a post which was quoted by another post in the same update, it would
immediately stop rendering all posts after hitting the post which
quoted the previous one.